add last vax type model
