+++ app/routes/model.$model.$dataset.$covariate.tsx
@@ -6,14 +6,21 @@ import ConfiguredPlot from "~/components/ConfiguredPlot";
 import {useLoaderData, useParams} from "@remix-run/react";
 import useSelectedModel from "~/hooks/useSelectedModel";
 
+interface Response {
+    status: "ERROR" | "SUCCESS"
+    data?: any
+    message?: string
+}
+
 export const loader = async ({params}) => {
     const jsonDirectory = "./data";
-    let data = {};
+    let data: Response;
     try {
         const fileContents = await fs.readFile([jsonDirectory, params.model, params.dataset, params.covariate, "res.json"].join("/"), "utf8");
-        data = JSON.parse(fileContents);
+        data = {status: "SUCCESS", data: JSON.parse(fileContents)};
     } catch (e) {
         console.error(e);
+        data = {status: "ERROR", message: "Data file not found"}
     }
     return json(
         data,
@@ -23,6 +30,7 @@ export const loader = async ({params}) => {
 export default function Index() {
     const {state, dispatch} = useContext<AppContext>(RootContext);
     const [status, selected] = useSelectedModel();
+    const data = useLoaderData<Response>();
     if (status == 404) {
         console.error(selected);
         return <div className={"text-center"}><h1>404</h1>
@@ -32,19 +40,13 @@ export default function Index() {
     if (!selected) {
         return null
     }
-    const {selectedModel} = selected;
-    if (status == 200) {
-        const data = useLoaderData<typeof loader>();
-        useEffect(() => {
-            const newState = {...state}
-            newState.selectedPlotOptions = Object.fromEntries(selectedModel.plots.map(p => [p.key,
-                Object.fromEntries(selectedModel.variables.concat([selectedModel.regressionModels[0]]).map(
-                    v => [v.key, "trace"]
-                ))
-            ]))
-            dispatch(newState);
-        }, []);
-
-        return data && selectedModel.plots.map(p => <ConfiguredPlot key={p.key} plot={p} data={data}></ConfiguredPlot>)
+    if (data.status == "ERROR") {
+        console.error("Data not found");
+        return <div className={"text-center"}><h1>404</h1>
+            <p>{data.message}</p>
+        </div>
     }
+    const {selectedModel} = selected;
+
+    return data.data && selectedModel.plots.map(p => <ConfiguredPlot key={p.key} plot={p} data={data.data}></ConfiguredPlot>)
 }
